+++ tests/test_proto_methods.rs
@@ -417,32 +417,6 @@ fn callable() {
     py_assert!(py, nc, "not callable(nc)");
 }
 
-#[allow(deprecated)]
-mod deprecated {
-    use super::*;
-
-    #[pyclass]
-    struct Callable;
-
-    #[pymethods]
-    impl Callable {
-        #[__call__]
-        fn __call__(&self, arg: i32) -> i32 {
-            arg * 6
-        }
-    }
-
-    #[test]
-    fn callable() {
-        let gil = Python::acquire_gil();
-        let py = gil.python();
-
-        let c = Py::new(py, Callable).unwrap();
-        py_assert!(py, c, "callable(c)");
-        py_assert!(py, c, "c(7) == 42");
-    }
-}
-
 #[pyclass]
 #[derive(Debug)]
 struct SetItem {